+++ app/controllers/services/configurations_controller.rb
@@ -4,22 +4,26 @@ class Services::ConfigurationsController < Services::ApplicationController
   before_action :ensure_in_session!
 
   def show
-    setup_show_variables!
-    @project_item = ProjectItem.new(session[session_key])
+    if ProjectItem::Wizard::OfferSelectionStep.new(session[session_key]).valid?
+      @step = step_class.new(session[session_key])
+    else
+      redirect_to service_offers_path(@service),
+                  alert: "Please select one of the service offer"
+    end
   end
 
   def update
-    @project_item = ProjectItem.new(configuration_params)
+    @step = step_class.new(configuration_params)
 
-    if @project_item.request_voucher
-      @project_item.voucher_id = ""
+    if @step.request_voucher
+      @step.voucher_id = ""
     end
 
-    if @project_item.valid?
-      session[session_key] = @project_item.attributes
+    if @step.valid?
+      save_in_session(@step)
       redirect_to service_summary_path(@service)
     else
-      setup_show_variables!
+      flash.now[:alert] = @step.error
       render :show
     end
   end
@@ -32,7 +36,7 @@ class Services::ConfigurationsController < Services::ApplicationController
           merge(status: :created)
     end
 
-    def setup_show_variables!
-      @projects = policy_scope(current_user.projects.active)
+    def step_class
+      ProjectItem::Wizard::ConfigurationStep
     end
 end